Change class name and add lombok annotation
