+++ backend/convertform/src/main/java/com/example/convertform/dto/input/AdGroupRecord.java
@@ -1,7 +1,9 @@
 package com.example.convertform.dto.input;
 
 import com.gh.mygreen.xlsmapper.annotation.XlsColumn;
+import lombok.Data;
 
+@Data
 public class AdGroupRecord {
     @XlsColumn(columnName = "Status")
     String status;
