+++ backend/convertform/src/main/java/com/example/convertform/dto/input/AreaRecord.java
@@ -1,7 +1,9 @@
 package com.example.convertform.dto.input;
 
 import com.gh.mygreen.xlsmapper.annotation.XlsColumn;
+import lombok.Data;
 
+@Data
 public class AreaRecord {
     @XlsColumn(columnName = "Campaign Name")
     String cName;
