+++ backend/convertform/src/main/java/com/example/convertform/dto/input/SearchTargetTable.java
@@ -1,7 +1,6 @@
 package com.example.convertform.dto.input;
 
-import com.example.convertform.dto.input.target_list_item.PlacementItem;
-import com.example.convertform.dto.input.target_list_item.SearchKeywordItem;
+import com.example.convertform.dto.input.target_list_item.SearchKeywordRecord;
 import com.gh.mygreen.xlsmapper.annotation.LabelledCellType;
 import com.gh.mygreen.xlsmapper.annotation.XlsHorizontalRecords;
 import com.gh.mygreen.xlsmapper.annotation.XlsLabelledCell;
@@ -11,9 +10,9 @@ import java.util.List;
 
 //Mỗi record là 1 table trong sheet
 @Data
-public class SearchTargetRecord {
+public class SearchTargetTable {
     @XlsHorizontalRecords(tableLabel = "Search Targeting", bottom = 3)
-    List<SearchKeywordItem> searchKeywordItems;
+    List<SearchKeywordRecord> searchKeywordRecords;
     @XlsLabelledCell(label = "Targeting Name", type = LabelledCellType.Right)
     String targetName;
     @XlsLabelledCell(label = "Search Keyword List Name", type = LabelledCellType.Right)
